chore(core): open confirm box before deleting permission scope (#4272)
Minor Quality of Life Improvement for Community Managers / Admins. Opens a Confirmation Box before a permission scope is deleted.

From past experiences "deleting" that scope doesn't actually irreversibly delete the configured permissions — unclear if this is intended behaviour or a bug. But erring on the side of caution I still choose a terminology which suggests that the action is destructive